requirements.txt
pytorch-lightning
torch
setuptools
lens-metric
unbabel-comet
ai2-olmo
vllm
openai
bert_score
gdown
nltk
sacremoses
flask
fairscale
tqdm
datasets
code-bert-score
sentence_transformers
filelock
sacrebleu
pyminifier
bootstrapped
accelerate
flash-attn
bitsandbytes
